formatting and added hydromt-hurrywave
