+++ cht_nesting/nest2/nest2_beware_in_hurrywavecht.py
@@ -1,7 +1,8 @@
-"""Nest 2 script for nesting BEWARE within hydromt_hurrywave HurrywaveModel.
+# -*- coding: utf-8 -*-
+"""
+Created on Fri Sep  3 13:40:56 2021
 
-Reads wave parameter time series from the HurrywaveModel's hurrywave_his.nc output
-and sets them as wave boundary conditions on the detail BEWARE model.
+This module defines the nest2_beware_in_hurrywave function for handling nesting of BEWARE models within HurryWave models.
 """
 
 import os
@@ -11,7 +12,7 @@ import pandas as pd
 import xarray as xr
 
 
-def nest2_beware_in_hurrywave(
+def nest2_beware_in_hurrywavecht(
     overall: Any,
     detail: Any,
     output_path: Optional[str] = None,
@@ -19,34 +20,33 @@ def nest2_beware_in_hurrywave(
     bc_path: Optional[str] = None,
     **kwargs,
 ) -> None:
-    """Nest a BEWARE model within a HurrywaveModel.
+    """
+    Nest a BEWARE model within a HurryWave model.
 
-    Parameters
-    ----------
-    overall : hydromt_hurrywave.HurrywaveModel
-        The coarse HurrywaveModel whose his output is read.
-    detail : BEWARE
-        The fine BEWARE model that receives wave boundary conditions.
-    output_path : str, optional
-        Directory containing the HurrywaveModel output files.
-    output_file : str, optional
-        Name of the his output file. Defaults to "hurrywave_his.nc".
-    bc_path : str, optional
-        If provided, write boundary conditions to disk.
+    Parameters:
+    overall (Any): The overall HurryWave model.
+    detail (Any): The detailed BEWARE model.
+    output_path (Optional[str]): The path to the output files. Default is None.
+    output_file (Optional[str]): The name of the output file. Default is None.
+    bc_path (Optional[str]): The path to the boundary conditions files. Default is None.
+    **kwargs: Additional keyword arguments.
     """
     if not output_path:
-        output_path = str(overall.root.path)
+        # Path of the overall output time series
+        output_path = overall.path
     if not output_file:
         output_file = "hurrywave_his.nc"
 
     file_name = os.path.join(output_path, output_file)
 
+    # Open netcdf file
     ddd = xr.open_dataset(file_name)
     stations = ddd.station_name.values
     all_stations = [str(st.strip())[2:-1] for st in stations]
 
     point_names = []
     if detail.wave_boundary_point:
+        # Find required boundary points
         for point in detail.wave_boundary_point:
             point_names.append(detail.name + "_" + point.name)
     else:
