+++ cht_nesting/nest2/nest2_hurrywave_in_hurrywave.py
@@ -1,13 +1,14 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Fri Sep  3 13:40:56 2021
+"""Nest 2 script for nesting hydromt_hurrywave HurrywaveModel within HurrywaveModel.
 
-This module defines the nest2_hurrywave_in_hurrywave function for handling nesting of HurryWave models within HurryWave models.
+Reads spectral output from the overall HurrywaveModel and sets spectral
+boundary conditions on the detail HurrywaveModel.
 """
 
 import os
-from typing import Any, Optional
+from typing import Any, List, Optional
 
+import numpy as np
+import pandas as pd
 import xarray as xr
 
 
@@ -20,66 +21,102 @@ def nest2_hurrywave_in_hurrywave(
     bc_path: Optional[str] = None,
     **kwargs,
 ) -> None:
-    """
-    Nest a HurryWave model within another HurryWave model.
+    """Nest a HurrywaveModel within another HurrywaveModel.
 
-    Parameters:
-    overall (Any): The overall HurryWave model.
-    detail (Any): The detailed HurryWave model.
-    obs_point_prefix (Optional[str]): The prefix for observation points. Default is None.
-    output_path (Optional[str]): The path to the output files. Default is None.
-    output_file (Optional[str]): The name of the output file. Default is None.
-    bc_path (Optional[str]): The path to the boundary conditions files. Default is None.
-    **kwargs: Additional keyword arguments.
+    Parameters
+    ----------
+    overall : hydromt_hurrywave.HurrywaveModel
+        The coarse HurrywaveModel whose spectral output is read.
+    detail : hydromt_hurrywave.HurrywaveModel
+        The fine HurrywaveModel that receives the spectral boundary conditions.
+    obs_point_prefix : str, optional
+        Prefix for observation point names.
+    output_path : str, optional
+        Directory containing the overall model's output files.
+    output_file : str, optional
+        Name of the spectral output file. Defaults to "hurrywave_sp2.nc".
+    bc_path : str, optional
+        If provided, write boundary conditions to disk.
     """
     if not output_path:
-        # Path of the overall output time series
-        output_path = overall.path
+        output_path = str(overall.root.path)
     if not output_file:
         output_file = "hurrywave_sp2.nc"
 
     file_name = os.path.join(output_path, output_file)
 
-    detail.boundary_conditions.forcing = "spectra"
+    # Read boundary points from detail model
+    bndfile = detail.config.get("bndfile")
+    if bndfile is None:
+        detail.config.set("bndfile", "hurrywave.bnd")
+    detail.boundary_conditions._read_boundary_points()
+    bnd_gdf = detail.boundary_conditions.gdf
 
-    # Open netcdf file
+    if len(bnd_gdf) == 0:
+        print("Warning: no boundary points found in detail model")
+        return
+
+    # Open the overall spectral output
     ddd = xr.open_dataset(file_name)
     stations = ddd.station_name.values
-    all_stations = [str(st.strip())[2:-1] for st in stations]
+    all_stations: List[str] = [
+        st.decode("utf-8").strip()
+        if isinstance(st, (bytes, np.bytes_))
+        else str(st).strip()
+        for st in stations
+    ]
 
-    if len(detail.boundary_conditions.gdf) == 0:
-        detail.input.variables.bndfile = "hurrywave.bnd"
-        detail.boundary_conditions.read_boundary_points()
+    # Build required point names
+    point_names: List[str] = [
+        f"{obs_point_prefix}_{row['name']}"
+        for _, row in bnd_gdf.iterrows()
+    ]
 
-    point_names = []
-    if len(detail.boundary_conditions.gdf) > 0:
-        for ind, row in detail.boundary_conditions.gdf.iterrows():
-            # Find required boundary points
-            point_names.append(obs_point_prefix + "_" + row["name"])
-    else:
-        point_names = all_stations.copy()
+    # Map point names to station indices
+    ireq: List[int] = []
+    for pname in point_names:
+        matched = False
+        for ist, st in enumerate(all_stations):
+            if pname.lower() == st.lower():
+                ireq.append(ist)
+                matched = True
+                break
+        if not matched:
+            raise ValueError(
+                f"Observation point '{pname}' not found in '{file_name}'. "
+                f"Available stations: {all_stations}"
+            )
 
     times = ddd.point_spectrum2d.coords["time"].values
     sigma = ddd.point_spectrum2d.coords["sigma"].values
     theta = ddd.point_spectrum2d.coords["theta"].values
 
-    ireq = []
-    for ip, point in enumerate(point_names):
-        for ist, st in enumerate(all_stations):
-            if point.lower() == st.lower():
-                ireq.append(ist)
-                break
+    # Extract spectra for required stations
+    sp2 = ddd.point_spectrum2d.values[:, ireq, :, :]  # (time, stations, theta, sigma)
 
-    for ind, row in detail.boundary_conditions.gdf.iterrows():
-        sp2 = ddd.point_spectrum2d.values[:, ireq[ind], :, :]
+    # Get boundary point coordinates
+    xs = np.array([row.geometry.x for _, row in bnd_gdf.iterrows()])
+    ys = np.array([row.geometry.y for _, row in bnd_gdf.iterrows()])
+    station_names = [row["name"] for _, row in bnd_gdf.iterrows()]
 
-        ds = xr.Dataset(
-            data_vars={"point_spectrum2d": (["time", "theta", "sigma"], sp2)},
-            coords={"time": times, "theta": theta, "sigma": sigma},
-        )
-        detail.boundary_conditions.gdf.loc[ind, "spectra"] = ds.to_array()
+    ddd.close()
+
+    # Build a proper spectra dataset for set_spectra()
+    ds_spectra = xr.Dataset(
+        data_vars={
+            "point_spectrum2d": (["time", "stations", "theta", "sigma"], sp2),
+            "station_x": ("stations", xs),
+            "station_y": ("stations", ys),
+        },
+        coords={
+            "time": times,
+            "stations": station_names,
+            "theta": theta,
+            "sigma": sigma,
+        },
+    )
+
+    detail.boundary_conditions.set_spectra(ds_spectra)
 
     if bc_path is not None:
-        detail.boundary_conditions.write_boundary_conditions_spectra(
-            file_name=os.path.join(bc_path, detail.input.variables.bspfile)
-        )
+        detail.boundary_conditions._write_boundary_spectra()
