+++ cht_nesting/nest2/nest2_sfincs_in_hurrywave.py
@@ -1,12 +1,11 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Fri Sep  3 13:40:56 2021
+"""Nest 2 script for nesting hydromt_sfincs SfincsModel within hydromt_hurrywave HurrywaveModel.
 
-This module defines the nest2_sfincs_in_hurrywave function for handling nesting of SFINCS models within HurryWave models.
+Reads wave parameter time series from the HurrywaveModel's hurrywave_his.nc output
+and sets them as SnapWave boundary conditions on the detail SfincsModel.
 """
 
 import os
-from typing import Any, Optional
+from typing import Any, List, Optional
 
 import numpy as np
 import pandas as pd
@@ -20,68 +19,88 @@ def nest2_sfincs_in_hurrywave(
     output_path: Optional[str] = None,
     output_file: Optional[str] = None,
     bc_path: Optional[str] = None,
-    **kwargs,
+    **kwargs: Any,
 ) -> None:
-    """
-    Nest a SFINCS model within a HurryWave model.
+    """Nest a SfincsModel within a HurrywaveModel (SnapWave BCs).
 
-    Parameters:
-    overall (Any): The overall HurryWave model.
-    detail (Any): The detailed SFINCS model.
-    obs_point_prefix (Optional[str]): The prefix for observation points. Default is None.
-    output_path (Optional[str]): The path to the output files. Default is None.
-    output_file (Optional[str]): The name of the output file. Default is None.
-    bc_path (Optional[str]): The path to the boundary conditions files. Default is None.
-    **kwargs: Additional keyword arguments.
+    Parameters
+    ----------
+    overall : hydromt_hurrywave.HurrywaveModel
+        The coarse HurrywaveModel whose his output is read.
+    detail : hydromt_sfincs.SfincsModel
+        The fine SfincsModel that receives the SnapWave boundary conditions.
+    obs_point_prefix : str, optional
+        Prefix for observation point names.
+    output_path : str, optional
+        Directory containing the HurrywaveModel output files.
+    output_file : str, optional
+        Name of the his output file. Defaults to "hurrywave_his.nc".
+    bc_path : str, optional
+        If provided, write boundary conditions to disk.
     """
     if not output_path:
-        # Path of the overall output time series
-        output_path = overall.path
+        output_path = str(overall.root.path)
     if not output_file:
         output_file = "hurrywave_his.nc"
 
-    file_name = os.path.join(output_path, output_file)
-    print("Nesting in " + file_name)
+    fn_his: str = os.path.join(output_path, output_file)
+    ddd = xr.open_dataset(fn_his)
 
-    # Open netcdf file
-    ddd = xr.open_dataset(file_name)
-    stations = ddd.station_name.values
-    all_stations = [str(st.strip())[2:-1] for st in stations]
+    # Decode station names
+    all_stations: List[str] = [
+        st.decode("utf-8").strip()
+        if isinstance(st, (bytes, np.bytes_))
+        else str(st).strip()
+        for st in ddd.station_name.values
+    ]
 
-    point_names = []
-    if len(detail.snapwave.boundary_conditions.gdf) > 0:
-        # Find required boundary points
-        for ind, point in detail.snapwave.boundary_conditions.gdf.iterrows():
-            point_names.append(obs_point_prefix + "_" + point["name"])
-    else:
-        point_names = all_stations.copy()
+    point_names: List[str] = [
+        f"{obs_point_prefix}_{str(ind + 1).zfill(4)}"
+        for ind, _row in detail.snapwave_boundary_conditions.gdf.iterrows()
+    ]
 
-    times = ddd.point_hm0.coords["time"].values
+    times = pd.DatetimeIndex(ddd.point_hm0.coords["time"].values)
 
-    ireq = []
-    for ip, point in enumerate(point_names):
+    ireq: List[int] = []
+    for pname in point_names:
+        matched = False
         for ist, st in enumerate(all_stations):
-            if point.lower() == st.lower():
+            if pname.lower() == st.lower():
                 ireq.append(ist)
+                matched = True
                 break
+        if not matched:
+            raise ValueError(
+                f"Observation point '{pname}' not found in '{fn_his}'. "
+                f"Available stations: {all_stations}"
+            )
 
-    for ip, point in detail.snapwave.boundary_conditions.gdf.iterrows():
-        hm0 = ddd.point_hm0.values[:, ireq[ip]]
-        tp = ddd.point_tp.values[:, ireq[ip]]
-        wavdir = ddd.point_wavdir.values[:, ireq[ip]]
-        dirspr = ddd.point_dirspr.values[:, ireq[ip]]
-        # Limit direction spreading to values between 4.0 and 55.0 degrees
-        dirspr = np.clip(dirspr, 4.0, 55.0)
-        # Limit period to values between 1.0 and 25.0 seconds
-        tp = np.clip(tp, 1.0, 25.0)
+    gdf_indices: List[Any] = [
+        ind for ind, _row in detail.snapwave_boundary_conditions.gdf.iterrows()
+    ]
 
-        df = pd.DataFrame(index=times)
-        df.insert(0, "hs", hm0)
-        df.insert(1, "tp", tp)
-        df.insert(2, "wd", wavdir)
-        df.insert(3, "ds", dirspr)
+    # Extract and clip wave parameters to valid physical ranges.
+    hm0_all: np.ndarray = ddd.point_hm0.values[:, ireq]
+    tp_all: np.ndarray = np.clip(ddd.point_tp.values[:, ireq], 1.0, 25.0)
+    wavdir_all: np.ndarray = ddd.point_wavdir.values[:, ireq]
+    dirspr_all: np.ndarray = np.clip(ddd.point_dirspr.values[:, ireq], 4.0, 55.0)
 
-        detail.snapwave.boundary_conditions.gdf.loc[ip, "timeseries"] = df
+    ddd.close()
+
+    ds_bcs = xr.Dataset(
+        data_vars={
+            "hs": (["time", "index"], hm0_all),
+            "tp": (["time", "index"], tp_all),
+            "wd": (["time", "index"], wavdir_all),
+            "ds": (["time", "index"], dirspr_all),
+        },
+        coords={
+            "time": times,
+            "index": gdf_indices,
+        },
+    )
+
+    detail.snapwave_boundary_conditions.set_timeseries(ds_bcs)
 
     if bc_path is not None:
-        detail.snapwave.boundary_conditions.write_boundary_conditions_timeseries()
+        detail.snapwave_boundary_conditions.write()
